LPD-33318 Refactor: move Playwright files to local folder
